fix not calling the converter
