+++ Neo4j.Driver/Neo4j.Driver.Tests.TestBackend/Protocol/Session/SessionRun.cs
@@ -3,38 +3,17 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 
 namespace Neo4j.Driver.Tests.TestBackend
 {
-	internal class SeessionRunTypeJsonConverter : JsonConverter<SessionRun.SessionRunType>
-	{
-		public override void WriteJson(JsonWriter writer, SessionRun.SessionRunType value, JsonSerializer serializer) =>
-			throw new NotImplementedException();
-
-		public override SessionRun.SessionRunType ReadJson(JsonReader reader, Type objectType, SessionRun.SessionRunType existingValue, bool hasExistingValue,
-			JsonSerializer serializer)
-		{
-			var json = JObject.Load(reader);
-			return new SessionRun.SessionRunType
-			{
-				sessionId = json["sessionId"]?.Value<string>(),
-				cypher = json["cypher"]?.Value<string>(),
-				parameters = json["params"]?.ToObject<Dictionary<string, CypherToNativeObject>>(),
-				txMeta = json["txMeta"]?.ToObject<Dictionary<string, object>>(),
-				timeout = json["timeout"]?.Value<int?>(),
-				TimeoutSet = json.ContainsKey("timeout")
-			};
-		}
-	}
-
-    internal class SessionRun : IProtocolObject
+	internal class SessionRun : IProtocolObject
     {
         public SessionRunType data { get; set; } = new SessionRunType();
 
         [JsonIgnore]
         private string ResultId { get; set; }
-
+		
+        [JsonConverter(typeof(SeessionRunTypeJsonConverter))]
         public class SessionRunType
         {
             public string sessionId { get; set; }
